pyproject.toml
[project]
name = "polymarket-arbitrage"
version = "3.0.0"
description = "Polymarket Arbitrage Detection System - Binary & NegRisk support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Quant Team"}
]
keywords = ["polymarket", "arbitrage", "trading", "prediction-markets", "crypto", "async"]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.13.2",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
trading = [
    "py-clob-client>=0.15.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
polyarb = "polyarb.main:main"
arb-scan = "polymarket_arbitrage_bot_v2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/polyarb"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
